change rust result to i64

+++ aocsuite/languages/templates/rust/exercise.rs
@@ -1,11 +1,12 @@
 use aocsuite::utils;
-fn exercise1(data: &str) -> i32 {
+fn exercise1(data: &str) -> i64 {
     return 0;
 }
-fn exercise2(data: &str) -> i32 {
+fn exercise2(data: &str) -> i64 {
     return 0;
 }
 
 fn main() {
     utils::run(exercise1, exercise2);
 }
+
